pyproject.toml
[project]
name = "fastapi-lib-edu"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncpg>=0.30.0",
    "fastapi[standard]>=0.115.12",
    "greenlet>=3.1.1",
    "pydantic-settings>=2.8.1",
    "sqlmodel>=0.0.24",
    "uvicorn>=0.34.0",
]
